Write a C program using C funtion

+++ Tutorial18.c
@@ -4,9 +4,9 @@
 int main ()
 {
     int a =3;
-    float b = (float )54/5;
+    float b = (54/5;
 
-    printf("The value of a is %f\n", b);
+    printf("The value of a is %f\n",(int) b);
     return 0;
 
-} 
+}
